requirements.txt
requests
BeautifulSoup
urllib3
dns.resolver
urllib.request
urllib.error
sys
threading
random
re